Refactor gen_keycode: Split generators (#61916)
* Split gen_keycode to make maintenance easier.
