+++ dev/tools/gen_keycodes/bin/gen_keycodes.dart
@@ -7,11 +7,19 @@ import 'dart:convert';
 import 'dart:io' hide Platform;
 
 import 'package:args/args.dart';
-import 'package:gen_keycodes/cc_code_gen.dart';
 import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as path;
 
-import 'package:gen_keycodes/dart_code_gen.dart';
+import 'package:gen_keycodes/android_code_gen.dart';
+import 'package:gen_keycodes/base_code_gen.dart';
+import 'package:gen_keycodes/macos_code_gen.dart';
+import 'package:gen_keycodes/fuchsia_code_gen.dart';
+import 'package:gen_keycodes/glfw_code_gen.dart';
+import 'package:gen_keycodes/gtk_code_gen.dart';
+import 'package:gen_keycodes/windows_code_gen.dart';
+import 'package:gen_keycodes/web_code_gen.dart';
+import 'package:gen_keycodes/keyboard_keys_code_gen.dart';
+import 'package:gen_keycodes/keyboard_maps_code_gen.dart';
 import 'package:gen_keycodes/key_data.dart';
 import 'package:gen_keycodes/utils.dart';
 
@@ -235,30 +243,49 @@ Future<void> main(List<String> rawArguments) async {
   if (!codeFile.existsSync()) {
     codeFile.createSync(recursive: true);
   }
+  print('Writing ${'key codes'.padRight(15)}${codeFile.absolute}');
+  await codeFile.writeAsString(KeyboardKeysCodeGenerator(data).generate());
 
   final File mapsFile = File(parsedArguments['maps'] as String);
   if (!mapsFile.existsSync()) {
     mapsFile.createSync(recursive: true);
   }
+  print('Writing ${'key maps'.padRight(15)}${mapsFile.absolute}');
+  await mapsFile.writeAsString(KeyboardMapsCodeGenerator(data).generate());
 
-  final DartCodeGenerator generator = DartCodeGenerator(data);
-  await codeFile.writeAsString(generator.generateKeyboardKeys());
-  await mapsFile.writeAsString(generator.generateKeyboardMaps());
+  for (final String platform in <String>['android', 'darwin', 'glfw', 'fuchsia', 'linux', 'windows', 'web']) {
+    PlatformCodeGenerator codeGenerator;
+    switch (platform) {
+      case 'glfw':
+        codeGenerator = GlfwCodeGenerator(data);
+        break;
+      case 'fuchsia':
+        codeGenerator = FuchsiaCodeGenerator(data);
+        break;
+      case 'android':
+        codeGenerator = AndroidCodeGenerator(data);
+        break;
+      case 'darwin':
+        codeGenerator = MacOsCodeGenerator(data);
+        break;
+      case 'windows':
+        codeGenerator = WindowsCodeGenerator(data);
+        break;
+      case 'linux':
+        codeGenerator = GtkCodeGenerator(data);
+        break;
+      case 'web':
+        codeGenerator = WebCodeGenerator(data);
+        break;
+      default:
+        assert(false);
+    }
 
-  final CcCodeGenerator ccCodeGenerator = CcCodeGenerator(data);
-  for (final String platform in <String>['android', 'darwin', 'glfw', 'gtk', 'fuchsia', 'linux', 'windows']) {
-    final File platformFile = File(path.join(flutterRoot.path, '..', path.join('engine', 'src', 'flutter', 'shell', 'platform', platform, 'keycodes', 'keyboard_map_$platform.h')));
+    final File platformFile = File(codeGenerator.outputPath(platform));
     if (!platformFile.existsSync()) {
       platformFile.createSync(recursive: true);
     }
-    print('Writing map ${platformFile.absolute}');
-    await platformFile.writeAsString(ccCodeGenerator.generateKeyboardMaps(platform));
+    print('Writing ${'$platform map'.padRight(15)}${platformFile.absolute}');
+    await platformFile.writeAsString(codeGenerator.generate());
   }
-
-  final File webPlatformFile = File(path.join(flutterRoot.path, '..', 'engine', 'src', 'flutter', path.join('lib', 'web_ui', 'lib', 'src', 'engine', 'keycodes', 'keyboard_map_web.dart')));
-  if (!webPlatformFile.existsSync()) {
-    webPlatformFile.createSync(recursive: true);
-  }
-  print('Writing map ${webPlatformFile.absolute}');
-  await webPlatformFile.writeAsString(generator.generateWebKeyboardMap());
 }
