--- dev/tools/gen_keycodes/data/keyboard_map_gtk_cc.tmpl
@@ -1,24 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#include <map>
-
-// DO NOT EDIT -- DO NOT EDIT -- DO NOT EDIT
-// This file is generated by flutter/flutter@dev/tools/gen_keycodes/bin/gen_keycodes.dart and
-// should not be edited directly.
-//
-// Edit the template dev/tools/gen_keycodes/data/keyboard_maps_gtk_cxx.tmpl instead.
-// See dev/tools/gen_keycodes/README.md for more information.
-
-/// Maps GTK-specific key codes to the matching [LogicalKeyboardKey].
-const std::map<int, int> g_gtk_to_logical_key = {
-@@@GTK_KEY_CODE_MAP@@@
-};
-
-/// A map of GTK key codes which have printable representations, but appear
-/// on the number pad. Used to provide different key objects for keys like
-/// KEY_EQUALS and NUMPAD_EQUALS.
-const std::map<int, int> g_gtk_numpad_map = {
-@@@GTK_NUMPAD_MAP@@@
-};
