+++ dev/tools/gen_keycodes/lib/keyboard_maps_code_gen.dart
@@ -2,37 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:io';
 import 'package:path/path.dart' as path;
 
-import 'package:gen_keycodes/key_data.dart';
-import 'package:gen_keycodes/utils.dart';
+import 'base_code_gen.dart';
+import 'key_data.dart';
+import 'utils.dart';
 
-/// Generates the keyboard_keys.dart and keyboard_maps.dart files, based on the
-/// information in the key data structure given to it.
-class CcCodeGenerator {
-  CcCodeGenerator(this.keyData);
 
-  /// Given an [input] string, wraps the text at 80 characters and prepends each
-  /// line with the [prefix] string. Use for generated comments.
-  String wrapString(String input, {String prefix = '  // '}) {
-    final int wrapWidth = 80 - prefix.length;
-    final StringBuffer result = StringBuffer();
-    final List<String> words = input.split(RegExp(r'\s+'));
-    String currentLine = words.removeAt(0);
-    for (final String word in words) {
-      if ((currentLine.length + word.length) < wrapWidth) {
-        currentLine += ' $word';
-      } else {
-        result.writeln('$prefix$currentLine');
-        currentLine = word;
-      }
-    }
-    if (currentLine.isNotEmpty) {
-      result.writeln('$prefix$currentLine');
-    }
-    return result.toString();
-  }
+/// Generates the keyboard_maps.dart files, based on the information in the key
+/// data structure given to it.
+class KeyboardMapsCodeGenerator extends BaseCodeGenerator {
+  KeyboardMapsCodeGenerator(KeyData keyData) : super(keyData);
 
   List<Key> get numpadKeyData {
     return keyData.data.where((Key entry) {
@@ -47,32 +27,13 @@ class CcCodeGenerator {
     }).toList();
   }
 
-  /// This generates the map of Flutter key codes to logical keys.
-  String get predefinedKeyCodeMap {
-    final StringBuffer keyCodeMap = StringBuffer();
-    for (final Key entry in keyData.data) {
-      keyCodeMap.writeln('    ${toHex(entry.flutterId, digits: 10)}: ${entry.constantName},');
-    }
-    for (final String entry in Key.synonyms.keys) {
-      // Use the first item in the synonyms as a template for the ID to use.
-      // It won't end up being the same value because it'll be in the pseudo-key
-      // plane.
-      final Key primaryKey = keyData.data.firstWhere((Key item) {
-        return item.name == Key.synonyms[entry][0];
-      }, orElse: () => null);
-      assert(primaryKey != null);
-      keyCodeMap.writeln('    ${toHex(Key.synonymPlane | primaryKey.flutterId, digits: 10)}: $entry,');
-    }
-    return keyCodeMap.toString().trimRight();
-  }
-
   /// This generates the map of GLFW number pad key codes to logical keys.
   String get glfwNumpadMap {
     final StringBuffer glfwNumpadMap = StringBuffer();
     for (final Key entry in numpadKeyData) {
       if (entry.glfwKeyCodes != null) {
         for (final int code in entry.glfwKeyCodes.cast<int>()) {
-          glfwNumpadMap.writeln('  { $code, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+          glfwNumpadMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
         }
       }
     }
@@ -85,19 +46,45 @@ class CcCodeGenerator {
     for (final Key entry in keyData.data) {
       if (entry.glfwKeyCodes != null) {
         for (final int code in entry.glfwKeyCodes.cast<int>()) {
-          glfwKeyCodeMap.writeln('  { $code, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+          glfwKeyCodeMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
         }
       }
     }
     return glfwKeyCodeMap.toString().trimRight();
   }
 
-  /// This generates the map of XKB scan codes to USB HID codes.
+  /// This generates the map of GTK number pad key codes to logical keys.
+  String get gtkNumpadMap {
+    final StringBuffer gtkNumpadMap = StringBuffer();
+    for (final Key entry in numpadKeyData) {
+      if (entry.gtkKeyCodes != null) {
+        for (final int code in entry.gtkKeyCodes.cast<int>()) {
+          gtkNumpadMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
+        }
+      }
+    }
+    return gtkNumpadMap.toString().trimRight();
+  }
+
+  /// This generates the map of GTK key codes to logical keys.
+  String get gtkKeyCodeMap {
+    final StringBuffer gtkKeyCodeMap = StringBuffer();
+    for (final Key entry in keyData.data) {
+      if (entry.gtkKeyCodes != null) {
+        for (final int code in entry.gtkKeyCodes.cast<int>()) {
+          gtkKeyCodeMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
+        }
+      }
+    }
+    return gtkKeyCodeMap.toString().trimRight();
+  }
+
+  /// This generates the map of XKB USB HID codes to physical keys.
   String get xkbScanCodeMap {
     final StringBuffer xkbScanCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.xKbScanCode != null) {
-        xkbScanCodeMap.writeln('  { ${toHex(entry.xKbScanCode)}, ${toHex(entry.usbHidCode)} },    // ${entry.constantName}');
+        xkbScanCodeMap.writeln('  ${toHex(entry.xKbScanCode)}: PhysicalKeyboardKey.${entry.constantName},');
       }
     }
     return xkbScanCodeMap.toString().trimRight();
@@ -109,7 +96,7 @@ class CcCodeGenerator {
     for (final Key entry in keyData.data) {
       if (entry.androidKeyCodes != null) {
         for (final int code in entry.androidKeyCodes.cast<int>()) {
-          androidKeyCodeMap.writeln('  { $code, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+          androidKeyCodeMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
         }
       }
     }
@@ -122,7 +109,7 @@ class CcCodeGenerator {
     for (final Key entry in numpadKeyData) {
       if (entry.androidKeyCodes != null) {
         for (final int code in entry.androidKeyCodes.cast<int>()) {
-          androidKeyCodeMap.writeln('  { $code, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+          androidKeyCodeMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
         }
       }
     }
@@ -135,7 +122,7 @@ class CcCodeGenerator {
     for (final Key entry in keyData.data) {
       if (entry.androidScanCodes != null) {
         for (final int code in entry.androidScanCodes.cast<int>()) {
-          androidScanCodeMap.writeln('  { $code, ${toHex(entry.usbHidCode)} },    // ${entry.constantName}');
+          androidScanCodeMap.writeln('  $code: PhysicalKeyboardKey.${entry.constantName},');
         }
       }
     }
@@ -147,7 +134,7 @@ class CcCodeGenerator {
     final StringBuffer windowsScanCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.windowsScanCode != null) {
-        windowsScanCodeMap.writeln('  { ${entry.windowsScanCode}, ${toHex(entry.usbHidCode)} },    // ${entry.constantName}');
+        windowsScanCodeMap.writeln('  ${toHex(entry.windowsScanCode)}: PhysicalKeyboardKey.${entry.constantName},');
       }
     }
     return windowsScanCodeMap.toString().trimRight();
@@ -157,20 +144,22 @@ class CcCodeGenerator {
   String get windowsNumpadMap {
     final StringBuffer windowsNumPadMap = StringBuffer();
     for (final Key entry in numpadKeyData) {
-      if (entry.windowsScanCode != null) {
-        windowsNumPadMap.writeln('  { ${toHex(entry.windowsScanCode)}, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+      if (entry.windowsKeyCodes != null){
+        for (final int code in entry.windowsKeyCodes) {
+          windowsNumPadMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
+        }
       }
     }
     return windowsNumPadMap.toString().trimRight();
   }
 
-  /// This generates the map of Android key codes to logical keys.
+  /// This generates the map of Windows key codes to logical keys.
   String get windowsKeyCodeMap {
     final StringBuffer windowsKeyCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.windowsKeyCodes != null) {
-        for (final int code in entry.windowsKeyCodes.cast<int>()) {
-          windowsKeyCodeMap.writeln('  { $code, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+        for (final int code in entry.windowsKeyCodes) {
+          windowsKeyCodeMap.writeln('  $code: LogicalKeyboardKey.${entry.constantName},');
         }
       }
     }
@@ -182,7 +171,7 @@ class CcCodeGenerator {
     final StringBuffer macOsScanCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.macOsScanCode != null) {
-        macOsScanCodeMap.writeln('  { ${toHex(entry.macOsScanCode)}, ${toHex(entry.usbHidCode)} },    // ${entry.constantName}');
+        macOsScanCodeMap.writeln('  ${toHex(entry.macOsScanCode)}: PhysicalKeyboardKey.${entry.constantName},');
       }
     }
     return macOsScanCodeMap.toString().trimRight();
@@ -193,7 +182,7 @@ class CcCodeGenerator {
     final StringBuffer macOsNumPadMap = StringBuffer();
     for (final Key entry in numpadKeyData) {
       if (entry.macOsScanCode != null) {
-        macOsNumPadMap.writeln('  { ${toHex(entry.macOsScanCode)}, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+        macOsNumPadMap.writeln('  ${toHex(entry.macOsScanCode)}: LogicalKeyboardKey.${entry.constantName},');
       }
     }
     return macOsNumPadMap.toString().trimRight();
@@ -203,7 +192,7 @@ class CcCodeGenerator {
     final StringBuffer macOsFunctionKeyMap = StringBuffer();
     for (final Key entry in functionKeyData) {
       if (entry.macOsScanCode != null) {
-        macOsFunctionKeyMap.writeln('  { ${toHex(entry.macOsScanCode)}, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+        macOsFunctionKeyMap.writeln('  ${toHex(entry.macOsScanCode)}: LogicalKeyboardKey.${entry.constantName},');
       }
     }
     return macOsFunctionKeyMap.toString().trimRight();
@@ -214,7 +203,7 @@ class CcCodeGenerator {
     final StringBuffer fuchsiaKeyCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.usbHidCode != null) {
-        fuchsiaKeyCodeMap.writeln('  { ${toHex(entry.flutterId)}, ${toHex(entry.flutterId, digits: 10)} },    // ${entry.constantName}');
+        fuchsiaKeyCodeMap.writeln('  ${toHex(entry.flutterId)}: LogicalKeyboardKey.${entry.constantName},');
       }
     }
     return fuchsiaKeyCodeMap.toString().trimRight();
@@ -225,7 +214,7 @@ class CcCodeGenerator {
     final StringBuffer fuchsiaScanCodeMap = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.usbHidCode != null) {
-        fuchsiaScanCodeMap.writeln(' { ${toHex(entry.usbHidCode)}, ${toHex(entry.usbHidCode)} },    // ${entry.constantName}');
+        fuchsiaScanCodeMap.writeln('  ${toHex(entry.usbHidCode)}: PhysicalKeyboardKey.${entry.constantName},');
       }
     }
     return fuchsiaScanCodeMap.toString().trimRight();
@@ -236,7 +225,7 @@ class CcCodeGenerator {
     final StringBuffer result = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.name != null) {
-        result.writeln("  '${entry.name}': ${toHex(entry.flutterId, digits: 10)},    // ${entry.constantName}");
+        result.writeln("  '${entry.name}': LogicalKeyboardKey.${entry.constantName},");
       }
     }
     return result.toString().trimRight();
@@ -247,7 +236,7 @@ class CcCodeGenerator {
     final StringBuffer result = StringBuffer();
     for (final Key entry in keyData.data) {
       if (entry.name != null) {
-        result.writeln("  '${entry.name}': ${toHex(entry.usbHidCode)},    // ${entry.constantName}");
+        result.writeln("  '${entry.name}': PhysicalKeyboardKey.${entry.constantName},");
       }
     }
     return result.toString().trimRight();
@@ -258,19 +247,18 @@ class CcCodeGenerator {
     final StringBuffer result = StringBuffer();
     for (final Key entry in numpadKeyData) {
       if (entry.name != null) {
-        result.writeln("  '${entry.name}': ${toHex(entry.flutterId, digits: 10)},    // ${entry.constantName}");
+        result.writeln("  '${entry.name}': LogicalKeyboardKey.${entry.constantName},");
       }
     }
     return result.toString().trimRight();
   }
 
-  /// Substitutes the various platform specific maps into the template file for
-  /// keyboard_maps.dart.
-  String generateKeyboardMaps(String platform) {
-    // There is no macOS keycode map since macOS uses keycode to represent a physical key.
-    // The LogicalKeyboardKey is generated by raw_keyboard_macos.dart from the unmodified characters
-    // from NSEvent.
-    final Map<String, String> mappings = <String, String>{
+  @override
+  String get templatePath => path.join(flutterRoot.path, 'dev', 'tools', 'gen_keycodes', 'data', 'keyboard_maps.tmpl');
+
+  @override
+  Map<String, String> mappings() {
+    return <String, String>{
       'ANDROID_SCAN_CODE_MAP': androidScanCodeMap,
       'ANDROID_KEY_CODE_MAP': androidKeyCodeMap,
       'ANDROID_NUMPAD_MAP': androidNumpadMap,
@@ -281,27 +269,15 @@ class CcCodeGenerator {
       'MACOS_FUNCTION_KEY_MAP': macOsFunctionKeyMap,
       'GLFW_KEY_CODE_MAP': glfwKeyCodeMap,
       'GLFW_NUMPAD_MAP': glfwNumpadMap,
+      'GTK_KEY_CODE_MAP': gtkKeyCodeMap,
+      'GTK_NUMPAD_MAP': gtkNumpadMap,
       'XKB_SCAN_CODE_MAP': xkbScanCodeMap,
       'WEB_LOGICAL_KEY_MAP': webLogicalKeyMap,
       'WEB_PHYSICAL_KEY_MAP': webPhysicalKeyMap,
       'WEB_NUMPAD_MAP': webNumpadMap,
-      'WINDOWS_SCAN_CODE_MAP': windowsScanCodeMap,
+      'WINDOWS_LOGICAL_KEY_MAP': windowsKeyCodeMap,
+      'WINDOWS_PHYSICAL_KEY_MAP': windowsScanCodeMap,
       'WINDOWS_NUMPAD_MAP': windowsNumpadMap,
-      'WINDOWS_KEY_CODE_MAP': windowsKeyCodeMap,
     };
-
-    final String template = File(path.join(flutterRoot.path, 'dev', 'tools', 'gen_keycodes', 'data', 'keyboard_map_${platform}_cc.tmpl')).readAsStringSync();
-    return _injectDictionary(template, mappings);
-  }
-
-  /// The database of keys loaded from disk.
-  final KeyData keyData;
-
-  static String _injectDictionary(String template, Map<String, String> dictionary) {
-    String result = template;
-    for (final String key in dictionary.keys) {
-      result = result.replaceAll('@@@$key@@@', dictionary[key]);
-    }
-    return result;
   }
 }
