+++ packages/flutter/lib/src/services/keyboard_maps.dart
@@ -636,6 +636,8 @@ const Map<int, LogicalKeyboardKey> kFuchsiaToLogicalKey = <int, LogicalKeyboardK
   0x1000c0073: LogicalKeyboardKey.brightnessMinimum,
   0x1000c0074: LogicalKeyboardKey.brightnessMaximum,
   0x1000c0075: LogicalKeyboardKey.brightnessAuto,
+  0x1000c0079: LogicalKeyboardKey.kbdIllumUp,
+  0x1000c007a: LogicalKeyboardKey.kbdIllumDown,
   0x1000c0083: LogicalKeyboardKey.mediaLast,
   0x1000c008c: LogicalKeyboardKey.launchPhone,
   0x1000c008d: LogicalKeyboardKey.programGuide,
@@ -908,6 +910,8 @@ const Map<int, PhysicalKeyboardKey> kFuchsiaToPhysicalKey = <int, PhysicalKeyboa
   0x000c0073: PhysicalKeyboardKey.brightnessMinimum,
   0x000c0074: PhysicalKeyboardKey.brightnessMaximum,
   0x000c0075: PhysicalKeyboardKey.brightnessAuto,
+  0x000c0079: PhysicalKeyboardKey.kbdIllumUp,
+  0x000c007a: PhysicalKeyboardKey.kbdIllumDown,
   0x000c0083: PhysicalKeyboardKey.mediaLast,
   0x000c008c: PhysicalKeyboardKey.launchPhone,
   0x000c008d: PhysicalKeyboardKey.programGuide,
@@ -1358,10 +1362,12 @@ const Map<int, LogicalKeyboardKey> kGtkToLogicalKey = <int, LogicalKeyboardKey>{
   57: LogicalKeyboardKey.digit9,
   48: LogicalKeyboardKey.digit0,
   65293: LogicalKeyboardKey.enter,
+  65076: LogicalKeyboardKey.enter,
   65307: LogicalKeyboardKey.escape,
   65288: LogicalKeyboardKey.backspace,
   65289: LogicalKeyboardKey.tab,
   65417: LogicalKeyboardKey.tab,
+  65056: LogicalKeyboardKey.tab,
   32: LogicalKeyboardKey.space,
   65408: LogicalKeyboardKey.space,
   45: LogicalKeyboardKey.minus,
@@ -1470,6 +1476,8 @@ const Map<int, LogicalKeyboardKey> kGtkToLogicalKey = <int, LogicalKeyboardKey>{
   65512: LogicalKeyboardKey.metaRight,
   269025026: LogicalKeyboardKey.brightnessUp,
   269025027: LogicalKeyboardKey.brightnessDown,
+  269025029: LogicalKeyboardKey.kbdIllumUp,
+  269025030: LogicalKeyboardKey.kbdIllumDown,
   269025134: LogicalKeyboardKey.launchPhone,
   269025044: LogicalKeyboardKey.mediaPlay,
   64790: LogicalKeyboardKey.mediaPlay,
@@ -1688,6 +1696,8 @@ const Map<int, PhysicalKeyboardKey> kLinuxToPhysicalKey = <int, PhysicalKeyboard
   0x00000258: PhysicalKeyboardKey.brightnessMinimum,
   0x00000259: PhysicalKeyboardKey.brightnessMaximum,
   0x000000fc: PhysicalKeyboardKey.brightnessAuto,
+  0x000000ee: PhysicalKeyboardKey.kbdIllumUp,
+  0x000000ed: PhysicalKeyboardKey.kbdIllumDown,
   0x0000019d: PhysicalKeyboardKey.mediaLast,
   0x000000b1: PhysicalKeyboardKey.launchPhone,
   0x00000172: PhysicalKeyboardKey.programGuide,
